Parsing

+++ LuceneTester.java
@@ -8,46 +8,23 @@ import java.util.Scanner;
 
 public class LuceneTester {
 public static void main(String[] args) throws Exception {
-	
-/*	String user = "root";
-    String password = "s3nsci0";
-    String host = "10.0.55.90";
-    int port=22;
-    ChannelSftp sftpChannel = null;
-    //String remoteFile="10.0.55.90/var/www/Library/Internal%20Document%20Repository/";
-
-    try{
-        JSch jsch = new JSch();
-        Session session = jsch.getSession(user, host, port);
-            session.setPassword(password);
-            session.setConfig("StrictHostKeyChecking", "no");
-        System.out.println("Establishing Connection...");
-        session.connect();
-            System.out.println("Connection established.");
-        System.out.println("Crating SFTP Channel.");
-        sftpChannel = (ChannelSftp) session.openChannel("sftp");
-        sftpChannel.connect();
-        System.out.println("SFTP Channel created.");
-        }
-    catch(Exception e){System.err.print(e);}
-*/
 
     //Input path of location for the index
-	//String indexDir = "/Users/Gina/Documents/OneDrive/txt_index";
-    String indexDir = "/var/www/library/index";
+	String indexDir = "/Users/Gina/Documents/OneDrive/txt_index";
+    //String indexDir = "/var/www/library/index";
 	//Input path of location for the directory that has all of the files
-	//String dataDir = "/Users/Gina/Documents/OneDrive/txt_data";
-    String dataDir = "/var/www/library/Internal Document Repository";
+	String dataDir = "/Users/Gina/Documents/OneDrive/txt_data";
+    //String dataDir = "/var/www/library/Internal Document Repository";
 	Indexer indexer = null;
 	   
 	File indexDirFile = new File(indexDir);
-	   
-	//Call clean class to clear all of the files created in the index directory from previous runs
 
+	System.out.println("Do you need to re-index the files? (y/n))");	
+	Scanner scan = new Scanner(System.in);
+	if (scan.nextLine().equals("y")){
 	clean.deleteFolderContents(indexDirFile);
-	   
-	//Call the Indexer.java file and create an indexer
 	Indexer.createIndex(indexDir, dataDir, indexer);
+	}
 	
 	//Ask for the string you are searching for in the files
 	 while (true) {
